requirements.txt
attrdict==2.0.1
imageio==2.6.1
matplotlib==3.1.1
pandas==0.25.3
numpy==1.16.4
Pillow==6.2.1
scikit-image==0.16.1
scipy==1.4.1
tqdm==4.36.1
yacs==0.1.6
tensorboard==2.1.1
joblib==0.14.0
motmetrics==1.1.3
moviepy==1.0.1
opencv-contrib-python==4.1.1.26
h5py==2.10.0
mazelib==0.9.4
